Move flaky test to a dedicated test
